+++ src/style/style_layer/overlap_mode.ts
