+++ src/symbol/collision_index.ts
@@ -15,7 +15,7 @@ import type {
     GlyphOffsetArray,
     SymbolLineVertexArray
 } from '../data/array_types.g';
-import type {OverlapMode} from '../style/style_layer/get_overlap_mode';
+import type {OverlapMode} from '../style/style_layer/overlap_mode';
 
 // When a symbol crosses the edge that causes it to be included in
 // collision detection, it will cause changes in the symbols around
